requirements.txt
beautifulsoup4
requests
flask
dotenv-dotenv